// models/create-endpoint-request-base.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bandwidth
 * Bandwidth\'s Communication APIs
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: letstalk@bandwidth.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EndpointDirectionEnum } from './endpoint-direction-enum';
// May contain unused imports in some cases
// @ts-ignore
import type { EndpointTypeEnum } from './endpoint-type-enum';

export interface CreateEndpointRequestBase {
    'type': EndpointTypeEnum;
    'direction': EndpointDirectionEnum;
    /**
     * The URL to send event callbacks to.
     */
    'eventCallbackUrl'?: string;
    /**
     * The URL to send event fallbacks to.
     */
    'eventFallbackUrl'?: string;
    /**
     * A tag for the endpoint.
     */
    'tag'?: string;
}